Change the package name "Controller" to "controller" and "usecase" to "useCases" for java name convention
