+++ src/main/java/useCases/impactor/MoneyImpactor.java
@@ -1,4 +1,4 @@
-package usecase.impactor;
+package useCases.impactor;
 
 import entity.Player;
